refactor scorer
